requirements.txt
# Project dependencies for intelligent-sre-mcp
httpx
fastapi
uvicorn[standard]
pydantic
kubernetes
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
